added owner history to auction details
