+++ components/TEMItems/TEMItemDetails/TEMItemDetails.module.css
@@ -1,7 +1,3 @@
-.clickable {
-    cursor: pointer;
-}
-
 .label {
     float: left;
     width: 250px;
