+++ components/TEMItems/TEMItemDetails/TEMItemDetails.tsx
@@ -5,6 +5,7 @@ import api from '../../../api/ApiHelper'
 import { convertTagToName, getStyleForTier, numberWithThousandsSeperators } from '../../../utils/Formatter'
 import { getLoadingElement } from '../../../utils/LoadingUtils'
 import { parsePlayer } from '../../../utils/Parser/APIResponseParser'
+import TEMOwnerHistory from '../TEMOwnerHistory/TEMOwnerHistory'
 import styles from './TEMItemDetails.module.css'
 
 interface Props {
@@ -13,44 +14,6 @@ interface Props {
 }
 
 function TEMItemDetails(props: Props) {
-    let [currentOwner, setCurrentOwner] = useState<Player>({})
-    let [previousOwners, setPreviousOwners] = useState<Player[]>([])
-    let [isLoading, setIsLoading] = useState(true)
-
-    useEffect(() => {
-        setIsLoading(true)
-
-        let promises = []
-
-        promises.push(
-            api.getPlayerName(props.detailEntry.currentOwner.playerUUID).then(name => {
-                setCurrentOwner(
-                    parsePlayer({
-                        name: name,
-                        uuid: props.detailEntry.currentOwner.playerUUID
-                    })
-                )
-            })
-        )
-
-        let prevOwners = []
-        props.detailEntry.previousOwners.forEach(previousOwner => {
-            api.getPlayerName(previousOwner.playerUUID).then(name => {
-                prevOwners.push(
-                    parsePlayer({
-                        name: name,
-                        uuid: previousOwner.playerUUID
-                    })
-                )
-            })
-        })
-
-        Promise.all(promises).then(() => {
-            setPreviousOwners(prevOwners)
-            setIsLoading(false)
-        })
-    }, [])
-
     function getDateString(date: Date) {
         if (!date) {
             return '-'
@@ -198,52 +161,13 @@ function TEMItemDetails(props: Props) {
         }
     }
 
-    function getPlayerElement(player: Player) {
-        return (
-            <>
-                <img
-                    crossOrigin="anonymous"
-                    className="playerHeadIcon"
-                    src={player.iconUrl}
-                    alt="player icon"
-                    height="36"
-                    style={{ marginRight: '10px' }}
-                    loading="lazy"
-                />
-                {player.name}
-            </>
-        )
-    }
-
     return (
         <>
             <Card style={{ marginBottom: '15px' }}>
                 <Card.Header>{getHeader()}</Card.Header>
-                <Card.Body>{isLoading ? getLoadingElement() : getInfoBody()}</Card.Body>
+                <Card.Body>{getInfoBody()}</Card.Body>
             </Card>
-            <Card>
-                <Card.Header>Owner History</Card.Header>
-                <Card.Body>
-                    <ul>
-                        <li>
-                            <Link href={`/player/${currentOwner.uuid}`}>
-                                <div className={styles.clickable}>
-                                    {getPlayerElement(currentOwner)} <span style={{ marginLeft: '10px', color: 'lime' }}>(Current)</span>
-                                </div>
-                            </Link>
-                        </li>
-                        {previousOwners.map(prevOwner => {
-                            return (
-                                <li>
-                                    <Link href={`/player/${prevOwner.uuid}`}>
-                                        <div className={styles.clickable}>{getPlayerElement(prevOwner)}</div>
-                                    </Link>
-                                </li>
-                            )
-                        })}
-                    </ul>
-                </Card.Body>
-            </Card>
+            <TEMOwnerHistory detailEntry={props.detailEntry} />
         </>
     )
 }
