+++ components/TEMItems/TEMOwnerHistory/TEMOwnerHistory.module.css
@@ -1,0 +1,3 @@
+.clickable {
+    cursor: pointer;
+}
